+++ src/subscriber.rs
@@ -61,7 +61,7 @@ where
 }
 
 impl<O, U: SubscriptionLike> SubscriptionLike for Subscriber<O, U> {
-  fn request(&mut self, requested: u128) {
+  fn request(&mut self, requested: usize) {
     self.subscription.request(requested);
   }
